notification-service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.1.5
MarkupSafe==3.0.2
multidict==6.4.3
propcache==0.3.1
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
twilio==9.5.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.1.5
MarkupSafe==3.0.2
multidict==6.4.3
packaging==24.2
propcache==0.3.1
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
twilio==9.5.2
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.19.0
